% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_status}
\alias{check_status}
\title{Check Request Status}
\usage{
check_status(request_obj)
}
\value{
If status code is not 200, returns a message to
console with appropriate error message and stops execution.
}
\description{
This function is called by other functions that submit a 
requests to revulytics API. It checks the status code
returned and tells the user if there was an error code and
exits the function.
}
\keyword{internal}
