% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Stop Function Quietly}
\usage{
stop_quietly()
}
\value{
Exits a function.
}
\description{
Quit a function execution without printing error messages. The
idea came from a Stack Overflow answer 
https://stackoverflow.com/questions/14469522/stop-an-r-program-without-error.
}
\keyword{internal}
